+++ src/scene/camera/PerspectiveCamera.js
@@ -5,6 +5,7 @@ import { Camera } from './Camera.js';
 const DEFAULT_FOVY = Math.PI / 3;
 
 export class PerspectiveCamera extends Camera {
+  
   constructor(fieldOfView = DEFAULT_FOVY, near = 0.1, far = 1000) {
     super(mat4.create(), mat4.create());
 
@@ -18,7 +19,9 @@ export class PerspectiveCamera extends Camera {
   /** @override */
   resize(viewportWidth = undefined, viewportHeight = undefined) {
     const aspectRatio =
-      typeof viewportWidth === 'undefined' ? 1 : viewportWidth / viewportHeight;
+      (typeof viewportWidth === 'undefined' || typeof viewportHeight === 'undefined')
+        ? 1
+        : viewportWidth / viewportHeight;
     const { near, far } = this.clippingPlane;
     mat4.perspective(
       this.projectionMatrix,